updated logger intrefaces
